requirements.txt
django==1.9.5
south==0.8.2
django-timezone-field==1.0
simplejson==3.3.0
